Implement GPS Location

+++ app/src/main/java/be/tabtabstudio/veganapp/data/entities/Location.java
@@ -1,10 +1,10 @@
 package be.tabtabstudio.veganapp.data.entities;
 
 public class Location {
-    public float lat;
-    public float lng;
+    public double lat;
+    public double lng;
 
-    public Location(float lat, float lng) {
+    public Location(double lat, double lng) {
         this.lat = lat;
         this.lng = lng;
     }
